refactor: UseCase 의존성 주입 모듈 분리 #156
