+++ app/src/main/java/com/kdjj/ratatouille/di/domain/DomainLayerModule.kt
@@ -1,10 +1,10 @@
 package com.kdjj.ratatouille.di.domain
 
-import com.kdjj.domain.di.UseCaseModule
+import com.kdjj.domain.di.ResultUseCaseModule
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 
-@Module(includes = [UseCaseModule::class])
+@Module(includes = [ResultUseCaseModule::class])
 @InstallIn(SingletonComponent::class)
 class DomainLayerModule
