+++ domain/src/main/java/com/kdjj/domain/di/FlowUseCaseModule.kt
@@ -1,0 +1,24 @@
+package com.kdjj.domain.di
+
+import com.kdjj.domain.model.Recipe
+import com.kdjj.domain.model.request.EmptyRequest
+import com.kdjj.domain.model.request.GetLocalRecipeFlowRequest
+import com.kdjj.domain.usecase.FlowUseCase
+import com.kdjj.domain.usecase.GetLocalRecipeFlowUseCase
+import com.kdjj.domain.usecase.GetRecipeUpdateFlowUseCase
+import dagger.Binds
+import dagger.Module
+
+@Module
+abstract class FlowUseCaseModule {
+
+    @Binds
+    internal abstract fun bindGetLocalRecipeFlowUseCase(
+        getLocalRecipeFlowUseCase: GetLocalRecipeFlowUseCase
+    ): FlowUseCase<GetLocalRecipeFlowRequest, Recipe>
+
+    @Binds
+    internal abstract fun bindGetRecipeUpdateFlowUseCase(
+        getRecipeUpdateFlowUseCase: GetRecipeUpdateFlowUseCase
+    ): FlowUseCase<EmptyRequest, Int>
+}
